+++ src/database/export/Exporter.js
@@ -1,8 +1,8 @@
-const GuildConfig = require('../config/GuildConfig');
-const ChannelConfig = require('../config/ChannelConfig');
-const Moderation = require('../database/Moderation.js');
-const AutoResponse = require('../database/AutoResponse.js');
-const BadWord = require('../database/BadWord.js');
+const GuildConfig = require('../../config/GuildConfig.js');
+const ChannelConfig = require('../../config/ChannelConfig.js');
+const Moderation = require('../Moderation.js');
+const AutoResponse = require('../AutoResponse.js');
+const BadWord = require('../BadWord.js');
 const {Snowflake} = require('discord.js');
 
 class Exporter {
